+++ webclient/src/store/reducers/board.js
@@ -17,9 +17,16 @@ export default (state = initialState, action) => {
     case actionTypes.ADD_LIST:
       newBoard.lists.push(action.list);
       return { ...state, board: newBoard };
-    case actionTypes.UPDATE_LIST:
-      newBoard.lists[action.list.index] = action.list;
+    case actionTypes.UPDATE_LIST: {
+      const listIndex = newBoard.lists.findIndex(
+        list => list._id === action.list._id
+      );
+      newBoard.lists[listIndex] = {
+        ...newBoard.lists[listIndex],
+        ...action.list
+      };
       return { ...state, board: newBoard };
+    }
     case actionTypes.ADD_CARD: {
       const listIndex = newBoard.lists.findIndex(
         list => list._id === action.newCard.listId
@@ -45,7 +52,6 @@ export default (state = initialState, action) => {
       return { ...state, board: newBoard };
 
     case actionTypes.CHANGE_CARD_TEXT: {
-      // props = listIndex, cardId, updatedCard
       // find list index
       const listIndex = state.board.lists.findIndex(
         list => list._id.toString() === action.listId
@@ -60,18 +66,30 @@ export default (state = initialState, action) => {
       };
       return { ...state, board: newBoard };
     }
-    // incomplete
-    case actionTypes.CHANGE_LIST:
-      const indexToSplice = newBoard.lists[action.prevList].cards.findIndex(
-        el => el.id === action.cardId
+    case actionTypes.CHANGE_LIST: {
+      // get the list indexes
+      // this is used because we (I) can't (can't think of a way to)
+      // modify arrays using just the fromList and toList ids
+      const from = newBoard.lists.findIndex(
+        list => list._id.toString() === action.fromList
       );
-      const todo = {
-        ...newBoard.lists[action.prevList].cards[indexToSplice],
-        inList: action.listToMoveTo
-      };
-      newBoard.lists[action.prevList].cards.splice(indexToSplice, 1);
-      newBoard.lists[action.listToMoveTo].cards.push(todo);
+      const to = newBoard.lists.findIndex(
+        list => list._id.toString() === action.toList
+      );
+      const cardIndex = newBoard.lists[from].cards.findIndex(
+        card => card._id.toString() === action.cardId
+      );
+      // copy the card to insert it in the new list
+      const card = { ...newBoard.lists[from].cards[cardIndex] };
+      // change the listId of the card (probably unnecessary property, not sure)
+      card.listId = action.toList;
+      // remove the card from the old list
+      newBoard.lists[from].cards.splice(cardIndex, 1);
+      // add the card to the new list
+      newBoard.lists[to].cards.push(card);
+      // return new board
       return { ...state, board: newBoard };
+    }
     default:
       return state;
   }